requirements.txt
# Core web framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# API and HTTP
requests>=2.31.0
python-multipart>=0.0.6

# Image processing
Pillow>=10.0.0
pillow-heif>=0.13.0
opencv-python>=4.8.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# AI/ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0

# Performance monitoring
psutil>=5.9.0
matplotlib>=3.7.0

# Development and utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
typing-extensions>=4.8.0

# Async support
aiofiles>=23.2.1

# Optional for enhanced features
scikit-learn>=1.3.0
scipy>=1.11.0
